fix test_bridge_groups
